Reorganize the components
