+++ packages/kbn-optimizer/limits.yml
@@ -42,6 +42,7 @@ pageLoadAssetSize:
   embeddable: 87309
   embeddableEnhanced: 22107
   enterpriseSearch: 66810
+  esqlDataGrid: 24582
   esUiShared: 326654
   eventAnnotation: 30000
   eventAnnotationListing: 25841
