+++ x-pack/plugins/search_inference_endpoints/public/inference_endpoints_router.tsx
@@ -10,7 +10,6 @@ import React from 'react';
 import { Redirect } from 'react-router-dom';
 import { InferenceEndpointsOverview } from './inference_endpoints_overview';
 
-import { InferenceEndpointsProvider } from './providers/inference_endpoints_provider';
 import { ROOT_PATH, SEARCH_INFERENCE_ENDPOINTS_PATH } from './routes';
 
 export const InferenceEndpointsRouter: React.FC = () => {
@@ -18,9 +17,7 @@ export const InferenceEndpointsRouter: React.FC = () => {
     <Routes>
       <Redirect exact from={ROOT_PATH} to={SEARCH_INFERENCE_ENDPOINTS_PATH} />
       <Route path={SEARCH_INFERENCE_ENDPOINTS_PATH}>
-        <InferenceEndpointsProvider>
-          <InferenceEndpointsOverview />
-        </InferenceEndpointsProvider>
+        <InferenceEndpointsOverview />
       </Route>
     </Routes>
   );
